Fix mobile view, move util functions, add dark theme mobile
