+++ src/components/header.tsx
@@ -1,5 +1,7 @@
 import React from 'react';
 import * as Icon from 'react-feather';
+import { changeTheme } from '../utils';
+
 export default function Header(props: { title: string }) {
   return (
     <div className='header'>
@@ -8,6 +10,9 @@ export default function Header(props: { title: string }) {
         <div className='search-box'>
           <Icon.Search />
           <input type='text' id='search' placeholder='Search' />
+        </div>
+        <div className='change-theme' onClick={changeTheme}>
+          <div className='icon-theme'></div>
         </div>
       </div>
     </div>
