+++ src/styles/header.scss
@@ -13,7 +13,10 @@
   .toolbar {
     display: flex;
     align-items: center;
-    margin-left: 1rem;
+    margin: 0 1rem;
+    @include sm {
+      justify-content: space-between;
+    }
     svg {
       position: relative;
       z-index: 3;
@@ -22,9 +25,6 @@
     .search-box:focus-within {
       svg {
         color: $primary;
-      }
-      #search {
-        width: 20rem;
       }
     }
     #search {
@@ -38,8 +38,37 @@
       border-radius: 12px;
       width: 15rem;
       transition: all 0.25s ease-in-out;
+      @include sm {
+        width: calc(100vw - 24px - 8rem);
+      }
       &:hover {
         background-color: darken($light, 3%);
+      }
+    }
+  }
+  .change-theme {
+    display: none;
+  }
+  @include sm {
+    .change-theme {
+      display: inherit;
+      .icon-theme {
+        background-color: $light;
+        &:active {
+          background-color: rgba($color: $primary, $alpha: 0.5);
+        }
+        &:hover {
+          background-color: darken($color: $light, $amount: 5%);
+        }
+        border-radius: 12px;
+        transition: all 0.1s ease-in-out;
+        width: 24px;
+        height: 24px;
+        padding: 11px;
+        background-size: 32px 32px;
+        background-position: center;
+        background-repeat: no-repeat;
+        background-image: url('/moon.svg');
       }
     }
   }
