+++ src/styles/sidebar.scss
@@ -133,6 +133,7 @@
       margin: 0;
       justify-content: space-between;
       padding: 0;
+      height: 100%;
     }
     a {
       display: inline-flex;
@@ -188,17 +189,21 @@
         flex-direction: column;
         align-items: center;
         justify-content: center;
-        padding: 1rem;
-        margin: -1.25rem 0 0 0;
+        padding: 0.5rem 1rem;
+        margin: -1.5rem 0 0 0;
         width: 40px;
-        background-color: rgba($color: #fff, $alpha: 0.9);
-
+        height: 100%;
+        box-shadow: 0px 15px 30px rgba(255, 0, 122, 0.24);
+        background-color: #fff;
         &.active {
-          background: rgba($color: $primary, $alpha: 0.9);
           color: #fff;
+          background: linear-gradient(90deg, $primary, lighten($primary, 20%));
         }
         .text {
-          font-size: 0.8rem;
+          font-size: 0.6rem;
+          letter-spacing: 1px;
+          margin-top: 4px;
+          font-weight: 600;
         }
         svg {
           width: 24px;
@@ -224,6 +229,9 @@
   }
 
   .change-theme {
+    @include sm {
+      display: none;
+    }
     margin-bottom: 2rem;
     .icon-theme {
       background-color: $light;
