+++ src/utils/index.tsx
@@ -1,0 +1,1 @@
+export { changeExpand, changeTheme } from './uiChanges';
